feat: moved server settings into its own component and added it to settings page
